# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Oona Developer Portal API
  version: 1.0.0
  description: Enterprise Service Catalog, IaC Pipeline Automation, and Compliance Portal for Oona Insurance.
servers:
  - url: http://127.0.0.1:8080
    description: Local Development Server
  - url: https://portal.oona-insurance.com
    description: Production Environment

paths:
  /health:
    get:
      summary: Basic liveness healthcheck
      tags: [System & Probes]
      responses:
        '200':
          description: Service alive
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"

  /live:
    get:
      summary: Kubernetes / Cloud Liveness Probe
      tags: [System & Probes]
      responses:
        '200':
          description: Process responsive
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "alive"

  /ready:
    get:
      summary: Kubernetes / Cloud Readiness Probe (checks DB pool and Valkey cache)
      tags: [System & Probes]
      responses:
        '200':
          description: Dependencies fully connected and ready
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ready"
                  database:
                    type: string
                    example: "connected"
                  valkey:
                    type: string
                    example: "connected"
        '503':
          description: Service unavailable (database or valkey down)

  /api/login:
    post:
      summary: User Login
      tags: [Authentication]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Login successful (Sets oona_token and oona_csrf cookies)
        '400':
          description: Missing or invalid credentials
        '429':
          description: Too Many Requests (Rate limit exceeded)

  /logout:
    post:
      summary: User Logout (Clears auth and CSRF cookies)
      tags: [Authentication]
      responses:
        '303':
          description: Redirect to login page

  /api/v1/tickets:
    post:
      summary: Create new service onboarding ticket
      tags: [Tickets]
      security:
        - CookieAuth: []
        - CSRFToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [domain, country, service_name, repo_url]
              properties:
                domain:
                  type: string
                  example: "integration"
                country:
                  type: string
                  example: "ph"
                service_name:
                  type: string
                  example: "health-renewal-svc"
                repo_url:
                  type: string
                  example: "https://github.com/oona-insurance/lmd-oona-ph-integration-health-renewal-svc"
                jira_issue_id:
                  type: string
                  example: "OONA-1042"
                description:
                  type: string
      responses:
        '201':
          description: Ticket created successfully
        '400':
          description: Invalid input format or missing fields
        '401':
          description: Unauthorized
        '403':
          description: CSRF validation failed

  /api/v1/tickets/{id}/approve:
    post:
      summary: Approve onboarding ticket & trigger Jenkins pipeline
      tags: [Approvals]
      security:
        - CookieAuth: []
        - CSRFToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ticket approved and provisioning queued
        '403':
          description: Forbidden (Requires role 'infra' or 'admin')

  /api/v1/tickets/{id}/reject:
    post:
      summary: Reject onboarding ticket with reason
      tags: [Approvals]
      security:
        - CookieAuth: []
        - CSRFToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Ticket rejected

components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: oona_token
    CSRFToken:
      type: apiKey
      in: header
      name: X-CSRF-Token